Conclusão de cadastro de motorista com json e rotas ok

+++ app/controllers/default.py
@@ -246,7 +246,7 @@ def cliente_add(email,senha,nomerazaosocial,foto,telefone,tipopessoa,cpfcnpj,log
         db.session.add(k)
         db.session.commit()
     else:
-        return {'sucesso':False, 'mensagem':'cliente existente.'}
+        return {'sucesso':False, 'mensagem':'cliente inexistente.'}
 
     return {'sucesso':True, 'mensagem':'cliente cadastrado com sucesso.'}#,'id_pessoa':k.id_pessoa,'cpfcnpj':i.cpfcnpj,'email':h.email}
 
@@ -310,19 +310,16 @@ def empresa_add(email,senha,nomerazaosocial,foto,telefone,tipopessoa,cpfcnpj,log
     if h == None:
         h = Usuario(email, senha)
         db.session.add(h)
-        #db.session.commit()
     
     i = Pessoa.query.filter(Pessoa.id_usuario == h.id_usuario).first()
     if i == None:
         i = Pessoa(nomerazaosocial,foto,telefone,h.id_usuario,tipopessoa,cpfcnpj)
         db.session.add(i)
-        #db.session.commit()
 
     j = Endereco.query.filter(Endereco.id_pessoa == i.id_pessoa).first()
     if j == None:
         j = Endereco(i.id_pessoa,logradouro,complemento,bairro,cidade,cep,uf)
         db.session.add(j)
-        #db.session.commit()
     
     k = Empresa.query.filter(Empresa.id_pessoa == i.id_pessoa).first()
     if k == None:
@@ -330,7 +327,7 @@ def empresa_add(email,senha,nomerazaosocial,foto,telefone,tipopessoa,cpfcnpj,log
         db.session.add(k)
         db.session.commit()
     else:
-        return {'sucesso':False, 'mensagem':'empresa existente.'}
+        return {'sucesso':False, 'mensagem':'empresa inexistente.'}
 
     return {'sucesso':True, 'mensagem':'empresa cadastrada com sucesso.'}#,'id_pessoa':k.id_pessoa,'cpfcnpj':i.cpfcnpj,'email':h.email}
 
@@ -344,17 +341,14 @@ def empresa_delete(id_pessoa):
     p = Endereco.query.get(d.id_pessoa)
 
     db.session.delete(d)
-    #db.session.commit()
     db.session.delete(p)
-    #db.session.commit()
     db.session.delete(e)
-    #db.session.commit()
     db.session.delete(f)
     db.session.commit()
 
     return {'sucesso':True, 'mensagem':'empresa removida com sucesso.'}
 
-def empresa(id_pessoa):
+def empresa_get(id_pessoa):
     g = Empresa.query.get(id_pessoa)
     if g == None:
         return {'sucesso':False, 'mensagem':'empresa não existe.'}
@@ -382,7 +376,7 @@ def empresa(id_pessoa):
         return jsonify(result), 400
 
     elif (request.method == 'GET'):
-        result = empresae_get(id_pessoa)
+        result = empresa_get(id_pessoa)
         if result['sucesso']:
             return jsonify(result), 200
         return jsonify(result), 400
@@ -446,56 +440,81 @@ def caminhao(id_caminhao):
 
 '''----------------------------Motorista--------------------------------'''
 
-def motorista_add(id_pessoa,id_caminhao,id_pessoa_emp):
-    i = Motorista(id_pessoa,id_caminhao,id_pessoa_emp)
-    db.session.add(i)
-    db.session.commit()
-    return "Motorista \"" + str(i.id_pessoa) + "\" incluido com sucesso!"
+def motorista_add(email,senha,nomerazaosocial,foto,telefone,tipopessoa,cpfcnpj,logradouro,complemento,bairro,cidade,cep,uf):
+    h = Usuario.query.filter(Usuario.email == email).first()
+    if h == None:
+        h = Usuario(email, senha)
+        db.session.add(h)
+    
+    i = Pessoa.query.filter(Pessoa.id_usuario == h.id_usuario).first()
+    if i == None:
+        i = Pessoa(nomerazaosocial,foto,telefone,h.id_usuario,tipopessoa,cpfcnpj)
+        db.session.add(i)
+
+    j = Endereco.query.filter(Endereco.id_pessoa == i.id_pessoa).first()
+    if j == None:
+        j = Endereco(i.id_pessoa,logradouro,complemento,bairro,cidade,cep,uf)
+        db.session.add(j)
+    
+    k = Motorista.query.filter(Motorista.id_pessoa == i.id_pessoa).first()
+    if k == None:
+        k = Motorista(i.id_pessoa,None,None)
+        db.session.add(k)
+        db.session.commit()
+    else:
+        return {'sucesso':False, 'mensagem':'motorista inexistente.'}
+
+    return {'sucesso':True, 'mensagem':'motorista cadastrado com sucesso.'}#,'id_pessoa':k.id_pessoa,'cpfcnpj':i.cpfcnpj,'email':h.email}
 
 def motorista_delete(id_pessoa):
     d = Motorista.query.get(id_pessoa)
+    if d == None:
+        return {'sucesso':False, 'mensagem':'motorista não existe.'}
+
+    e = Pessoa.query.get(d.id_pessoa)
+    f = Usuario.query.get(e.id_usuario)
+    p = Endereco.query.get(d.id_pessoa)
+
     db.session.delete(d)
+    db.session.delete(p)
+    db.session.delete(e)
+    db.session.delete(f)
     db.session.commit()
-    return "Motorista \"" + str(d.id_pessoa) + "\" excluído com sucesso!"
+
+    return {'sucesso':True, 'mensagem':'motorista removido com sucesso.'}
 
 def motorista_get(id_pessoa):
     g = Motorista.query.get(id_pessoa)
-    return "Motorista \"" + str(g.id_pessoa)
+    if g == None:
+        return {'sucesso':False, 'mensagem':'motorista não existe.'}
+    h = Pessoa.query.get(str(g.id_pessoa))
+    i = Usuario.query.get(h.id_usuario)
 
-def motorista_update(id_pessoa,id_caminhao,id_pessoa_emp):
-    u = Motorista.query.get(id_pessoa)
-    u.id_caminhao = id_caminhao
-    u.id_pessoa_emp = id_pessoa_emp
-    db.session.commit()
-    return "Motorista \"" + str(u.id_pessoa) + "\" alterado com sucesso!"
+    return {'sucesso':True,'mensagem':'motorista retornado com sucesso.','id_pessoa':g.id_pessoa,'cpfcnpj':h.cpfcnpj,'nomerazaosocial':h.nomerazaosocial,'email':i.email}
 
-@app.route("/motorista/<id_pessoa>",methods=['GET'])
+@app.route("/motorista/<id_pessoa>",methods=['GET','DELETE'])
 @app.route("/motorista/", defaults={'id_pessoa': None}, methods=['POST','GET','DELETE','PUT'])
 @app.route("/motorista", defaults={'id_pessoa': None}, methods=['POST','GET','DELETE','PUT'])
 def motorista(id_pessoa):
     if (request.method == 'POST'):
         some_json = request.get_json()
-        if motorista_add(some_json['id_pessoa'],some_json['id_caminhao'],some_json['id_pessoa_emp']):
-            return jsonify({'sucesso':True}), 201
-        return jsonify({'sucesso':False}), 400
+        result = motorista_add(some_json['email'],some_json['senha'],some_json['nomerazaosocial'],some_json['foto'],some_json['telefone'],some_json['tipopessoa'],some_json['cpfcnpj'],some_json['logradouro'],some_json['complemento'],some_json['bairro'],some_json['cidade'],some_json['cep'],some_json['uf'])
+        if result['sucesso']:
+            return jsonify(result), 201
+        return jsonify(result), 400
 
     elif (request.method == 'DELETE'):
         some_json = request.get_json()
-        if motorista_delete(some_json['id_pessoa']):
-            return jsonify({'sucesso':True}), 202
-        return jsonify({'sucesso':False}), 400
+        result = motorista_delete(id_pessoa)
+        if result['sucesso']:
+            return jsonify(result), 202
+        return jsonify(result), 400
 
     elif (request.method == 'GET'):
-        result =motorista_get(id_pessoa)
-        if result:
-            return jsonify({'sucesso':True, 'id_pessoa':result['id_pessoa'],'nomerazaosocial':result['nomerazaosocial'],'telefone':result['telefone'],'tipopessoa':result['tipopessoa'],'cpfcnpj':result['cpfcnpj']})
-        return jsonify({'sucesso':False})
-    
-    elif (request.method == 'PUT'):
-        some_json = request.get_json()
-        if motorista_update(some_json['id_pessoa'],some_json['id_caminhao'],some_json['id_pessoa_emp']):
-            return jsonify({'sucesso':True}), 200
-        return jsonify({'sucesso':False}), 400
+        result = motorista_get(id_pessoa)
+        if result['sucesso']:
+            return jsonify(result), 200
+        return jsonify(result), 400
 
 
 '''----------------------------Pedido--------------------------------'''
